+++ api-tests/tests/pg/setup/backup-storage.setup.ts
@@ -1,5 +1,6 @@
 // everest
 // Copyright (C) 2023 Percona LLC
+// Copyright (C) 2026 The OpenEverest Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -38,6 +39,6 @@ setup.describe.serial('PG Backup Storage setup', () => {
       verifyTLS: false,
       forcePathStyle: true,
     }
-    await th.createBackupStorageWithData(request, payload)
+    await th.createBackupStorageWithDataV1(request, payload)
   });
 });
