+++ api-tests/tests/utils/api.ts
@@ -231,7 +231,7 @@ export const deleteDBClusterRaw = async (request, name) => {
 }
 
 // --------------------- Backup Storage helpers -----------------------------------------
-export const getBackupStorageS3Payload = (bsName: string) => {
+export const getBackupStorageS3PayloadV1 = (bsName: string) => {
   const payload = {
     type: 's3',
     name: bsName,
@@ -246,7 +246,7 @@ export const getBackupStorageS3Payload = (bsName: string) => {
   return JSON.parse(JSON.stringify(payload))
 }
 
-export const getBackupStorageAzurePayload = (bsName: string) => {
+export const getBackupStorageAzurePayloadV1 = (bsName: string) => {
   const payload = {
     type: 'azure',
     name: bsName,
@@ -260,56 +260,56 @@ export const getBackupStorageAzurePayload = (bsName: string) => {
   return JSON.parse(JSON.stringify(payload))
 }
 
-export const createBackupStorageS3 = async (request, name) => {
-  const payload = getBackupStorageS3Payload(name)
-  return await createBackupStorageWithData(request, payload)
+export const createBackupStorageS3V1 = async (request, name) => {
+  const payload = getBackupStorageS3PayloadV1(name)
+  return await createBackupStorageWithDataV1(request, payload)
 }
 
-export const createBackupStorageWithData = async (request, data) => {
-  const response = await createBackupStorageWithDataRaw(request, data)
+export const createBackupStorageWithDataV1 = async (request, data) => {
+  const response = await createBackupStorageWithDataRawV1(request, data)
   await checkError(response)
   return (await response.json())
 }
 
-export const createBackupStorageWithDataRaw = async (request, data) => {
+export const createBackupStorageWithDataRawV1 = async (request, data) => {
   return await request.post(`/v1/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages`, {data: data})
 }
 
-export const getBackupStorage = async (request, name) => {
-  const response = await getBackupStorageRaw(request, name)
+export const getBackupStorageV1 = async (request, name) => {
+  const response = await getBackupStorageRawV1(request, name)
   await checkError(response)
   return (await response.json())
 }
 
-export const getBackupStorageRaw = async (request, name) => {
+export const getBackupStorageRawV1 = async (request, name) => {
   return await request.get(`/v1/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`)
 }
 
-export const listBackupStorages = async (request) => {
-  const response = await listBackupStoragesRaw(request)
+export const listBackupStoragesV1 = async (request) => {
+  const response = await listBackupStoragesRawV1(request)
   await checkError(response)
   return (await response.json())
 }
 
-export const listBackupStoragesRaw = async (request) => {
+export const listBackupStoragesRawV1 = async (request) => {
   return await request.get(`/v1/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages`)
 }
 
-export const updateBackupStorage = async (request, name, data) => {
-  const response = await updateBackupStorageRaw(request, name, data)
+export const updateBackupStorageV1 = async (request, name, data) => {
+  const response = await updateBackupStorageRawV1(request, name, data)
   await checkError(response)
   return (await response.json())
 }
 
-export const updateBackupStorageRaw = async (request, name, data) => {
+export const updateBackupStorageRawV1 = async (request, name, data) => {
   return await request.patch(`/v1/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`, {data: data})
 }
 
-export const deleteBackupStorage = async (request, name) => {
+export const deleteBackupStorageV1 = async (request, name) => {
   // Wait for deletion mark.
   await expect(async () => {
-    await deleteBackupStorageRaw(request, name)
-    const res = await getBackupStorageRaw(request, name)
+    await deleteBackupStorageRawV1(request, name)
+    const res = await getBackupStorageRawV1(request, name)
     await checkResourceDeletion(res)
   }).toPass({
     intervals: [1000],
@@ -317,7 +317,7 @@ export const deleteBackupStorage = async (request, name) => {
   })
 }
 
-export const deleteBackupStorageRaw = async (request, name) => {
+export const deleteBackupStorageRawV1 = async (request, name) => {
   return await request.delete(`/v1/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`)
 }
 
@@ -325,7 +325,7 @@ export const deleteBackupStorageRaw = async (request, name) => {
 
 const CLUSTER_NAME = 'main'
 
-export const getBackupStorageV2Payload = (bsName: string) => {
+export const getBackupStoragePayload = (bsName: string) => {
   return {
     metadata: {
       name: bsName,
@@ -344,51 +344,51 @@ export const getBackupStorageV2Payload = (bsName: string) => {
   }
 }
 
-export const createBackupStorageV2 = async (request, data) => {
-  const response = await createBackupStorageRawV2(request, data)
+export const generateBackupStorage = async (request, data) => {
+  const response = await createBackupStorageRaw(request, data)
   await checkError(response)
   return (await response.json())
 }
 
-export const createBackupStorageRawV2 = async (request, data) => {
+export const createBackupStorageRaw = async (request, data) => {
   return await request.post(`/v1/clusters/${CLUSTER_NAME}/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages`, {data: data})
 }
 
-export const getBackupStorageV2 = async (request, name) => {
-  const response = await getBackupStorageRawV2(request, name)
+export const getBackupStorage = async (request, name) => {
+  const response = await getBackupStorageRaw(request, name)
   await checkError(response)
   return (await response.json())
 }
 
-export const getBackupStorageRawV2 = async (request, name) => {
+export const getBackupStorageRaw = async (request, name) => {
   return await request.get(`/v1/clusters/${CLUSTER_NAME}/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`)
 }
 
-export const listBackupStoragesV2 = async (request) => {
-  const response = await listBackupStoragesRawV2(request)
+export const listBackupStorages = async (request) => {
+  const response = await listBackupStoragesRaw(request)
   await checkError(response)
   return (await response.json())
 }
 
-export const listBackupStoragesRawV2 = async (request) => {
+export const listBackupStoragesRaw = async (request) => {
   return await request.get(`/v1/clusters/${CLUSTER_NAME}/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages`)
 }
 
-export const updateBackupStorageV2 = async (request, name, data) => {
-  const response = await updateBackupStorageRawV2(request, name, data)
+export const updateBackupStorage = async (request, name, data) => {
+  const response = await updateBackupStorageRaw(request, name, data)
   await checkError(response)
   return (await response.json())
 }
 
-export const updateBackupStorageRawV2 = async (request, name, data) => {
+export const updateBackupStorageRaw = async (request, name, data) => {
   return await request.put(`/v1/clusters/${CLUSTER_NAME}/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`, {data: data})
 }
 
-export const deleteBackupStorageV2 = async (request, name) => {
+export const deleteBackupStorage = async (request, name) => {
   // Wait for deletion mark.
   await expect(async () => {
-    await deleteBackupStorageRawV2(request, name)
-    const res = await getBackupStorageRawV2(request, name)
+    await deleteBackupStorageRaw(request, name)
+    const res = await getBackupStorageRaw(request, name)
     await checkResourceDeletion(res)
   }).toPass({
     intervals: [1000],
@@ -396,7 +396,7 @@ export const deleteBackupStorageV2 = async (request, name) => {
   })
 }
 
-export const deleteBackupStorageRawV2 = async (request, name) => {
+export const deleteBackupStorageRaw = async (request, name) => {
   return await request.delete(`/v1/clusters/${CLUSTER_NAME}/namespaces/${EVEREST_CI_NAMESPACE}/backup-storages/${name}`)
 }
